+++ src/information/information.jl
@@ -5,3 +5,7 @@ include("information_measures.jl")
 include("information_functions.jl")
 include("information_estimators/information_estimators.jl")
 include("information_definitions/information_definitions.jl")
+
+# Specific estimators must be included after definitions.
+
+include("information_estimators/mutual_info_estimators/mutual_info_estimators.jl")