+++ src/information/information_definitions/conditional_mutual_informations/conditional_mutual_informations.jl
@@ -5,7 +5,7 @@ max_inputs_vars(::ConditionalMutualInformation) = 3
 
 
 # Generic H4-formulation of CMI
-function marginal_entropies_cmi4h_differential(est::DifferentialDecomposition{<:ConditionalMutualInformation, <:DifferentialInfoEstimator}, x, y, z)
+function marginal_entropies_cmi4h_differential(est::EntropyDecomposition{<:ConditionalMutualInformation, <:DifferentialInfoEstimator}, x, y, z)
     Z = StateSpaceSet(z)
     Y = StateSpaceSet(y)
     X = StateSpaceSet(x)
@@ -13,28 +13,31 @@ function marginal_entropies_cmi4h_differential(est::DifferentialDecomposition{<:
     YZ = StateSpaceSet(Y, Z)
     XYZ = StateSpaceSet(X, Y, Z)
 
-    HXZ = entropy(est.est, XZ)
-    HYZ = entropy(est.est, YZ)
-    HXYZ = entropy(est.est, XYZ)
-    HZ = entropy(est.est, Z)
+    modified_est = estimator_with_overridden_parameters(est.definition, est.est)
+    HXZ = entropy(modified_est, XZ)
+    HYZ = entropy(modified_est, YZ)
+    HXYZ = entropy(modified_est, XYZ)
+    HZ = entropy(modified_est, Z)
 
     return HXZ, HYZ, HXYZ, HZ
 end
 
-function marginal_entropies_cmi4h_discrete(est::DifferentialDecomposition{<:ConditionalMutualInformation, <:DifferentialInfoEstimator}, x, y, z)
+function marginal_entropies_cmi4h_discrete(est::EntropyDecomposition{<:ConditionalMutualInformation, <:DiscreteInfoEstimator}, x, y, z)
     # Encode marginals to integers based on the outcome space.
-    eX, eY, eZ = codified_marginals(est, x, y, z)
+    eX, eY, eZ = codified_marginals(est.discretization, x, y, z)
     eXZ = StateSpaceSet(eX, eZ)
     eYZ = StateSpaceSet(eY, eZ)
     eXYZ = StateSpaceSet(eX, eY, eZ)
  
     # The outcome space is no longer relevant from this point on. We're done discretizing, 
-    # so now we can just count (i.e. use `UniqueElements`).
+    # so now we can just count (i.e. use `UniqueElements` as the outcome space).
     o = UniqueElements()
-    HXZ = information(est.mest, est.pest, o, eXZ)
-    HYZ = information(est.mest, est.pest, o, eYZ)
-    HXYZ = information(est.mest, est.pest, o, eXYZ)
-    HZ = information(est.mest, est.pest, o, eZ)
+
+    modified_est = estimator_with_overridden_parameters(est.definition, est.mest)
+    HXZ = information(modified_est, est.pest, o, eXZ)
+    HYZ = information(modified_est, est.pest, o, eYZ)
+    HXYZ = information(modified_est, est.pest, o, eXYZ)
+    HZ = information(modified_est, est.pest, o, eZ)
 
     return HXZ, HYZ, HXYZ, HZ
 end