+++ src/information/information_definitions/information_definitions.jl
@@ -1,3 +1,4 @@
+
 include("divergences_and_distances/divergences_and_distances.jl")
 include("joint_entropies/joint_entropies.jl")
 include("conditional_entropies/conditional_entropies.jl")
@@ -5,3 +6,5 @@ include("mutual_informations/mutual_informations.jl")
 include("conditional_mutual_informations/conditional_mutual_informations.jl")
 
 include("partial_mutual_information.jl")
+
+include("override_parameters.jl")