+++ src/information/information_definitions/joint_entropies/JointEntropyTsallis.jl
@@ -14,25 +14,25 @@ Given two two discrete random variables ``X`` and ``Y`` with ranges ``\\mathcal{
 ``\\mathcal{X}``, [Furuichi2006](@citet) defines the Tsallis joint entropy as
 
 ```math
-H_q^T(X, Y) = -\\sum_{x\\in \\mathcal{X}, y \\in \\mathcal{Y}} p(x, y)^q \\log_q p(x, y)
+H_q^T(X, Y) = -\\sum_{x\\in \\mathcal{X}, y \\in \\mathcal{Y}} p(x, y)^q \\log_q p(x, y),
 ```
 
-where ``log_q(x, q) = \\dfrac{x^{1-q} - 1}{1-q}`` is the q-logarithm.
+where ``log_q(x, q) = \\dfrac{x^{1-q} - 1}{1-q}`` is the q-logarithm, and 
+we define ``log_q(x, q) := 0`` if ``q = 0``.
 """
-struct JointEntropyTsallis{E<:Tsallis} <: JointEntropy
-    e::E
-    function JointEntropyTsallis(; base = 2, q = 1.5)
-        e = Tsallis(; base, q)
-        new{typeof(e)}(e)
-    end
+Base.@kwdef struct JointEntropyTsallis{B, Q} <: JointEntropy
+    base::B = 2
+    q::Q = 1.5
 end
 
 function information(definition::JointEntropyTsallis, pxy::Probabilities{T, 2}) where T
-    base = definition.e.base
-    q = definition.e.q
+    (; base, q) = definition
+    
     h = 0.0
     for p in pxy
-        h += p^q * logq(p, q)
+        if p != 0.0 # Define logq(0) = 0
+            h += p^q * logq(p, q)
+        end
     end
     h = -h
     return _convert_logunit(h, ℯ, base)