+++ src/information/information_estimators/information_estimators.jl
@@ -1,3 +1,9 @@
 include("Contingency.jl")
 include("DiscreteDecomposition.jl")
 include("DifferentialDecomposition.jl")
+
+include("EntropyDecomposition.jl")
+include("MIDecomposition.jl")
+
+abstract type MutualInformationEstimator{M} <: BivariateInformationMeasureEstimator end
+abstract type ConditionalMutualInformationEstimator{M} <: MultivariateInformationMeasureEstimator end