+++ test/information/information.jl
@@ -6,3 +6,5 @@ include("conditional_mutual_informations/conditional_mutual_informations.jl")
 
 # Estimators of the information measures.
 include("estimators/discrete/joint.jl")
+
+include("internal_api.jl")